+++ src/plotery.js
@@ -17,8 +17,8 @@ export { AngularAxis } from './polar/angular-axis.component';
 export { PolarLine } from './polar/polar-line.component';
 export { PolarSector } from './polar/polar-sector.component';
 
-export { Zoom } from './interactions/zoom.component';
 export { Pointer } from './interactions/pointer.component';
+export { BoxZoom } from './interactions/box-zoom.component';
 
 export { linear } from './math/interpolations/linear';
 export { cardinal } from './math/interpolations/cardinal';