+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-There are no changes yet.
+### Fixed
+
+- Fix axis labels positions.
 
 
 ## [0.3.0] - 2020-06-13